Name change and sub impl.
Change the name of the class to BigInt to better reflect the type,

as it will support big negative numbers also with BigUInt to take the role of the unsigned int.

Implement the sub operation for BigInt, the implementation is not complete,

as it does not support negative numbers and the tests are not as extensive as they should be.